Added in some basic tests.
It's a bit tricky to test functions that print to STDOUT, but I've included a listener class if you want to experiment with it. When testing code that has expected output or return values you would use `XCTAssert` methods in order to test for success and failure conditions.

Note with the basic tests here we are only testing the `Functions` file, but that gives us coverage of the `Math` file as well. Generally you are only going to test your public APIs. You can use code coverage values to see that all of the expected code is actually being run.